Move Event and FloatWriter to 'property'
