+++ src/bms/player/beatoraja/skin/SkinObject.java
@@ -2,8 +2,6 @@ package bms.player.beatoraja.skin;
 
 import bms.player.beatoraja.MainState;
 import bms.player.beatoraja.skin.Skin.SkinObjectRenderer;
-import bms.player.beatoraja.skin.event.Event;
-import bms.player.beatoraja.skin.event.EventFactory;
 import bms.player.beatoraja.skin.property.*;
 
 import com.badlogic.gdx.graphics.Color;
